+++ lib/postgres/index.d.ts
@@ -4,6 +4,6 @@ declare type Props = {
     eventSchema?: string;
     poolConfig: PoolConfig;
 };
-export declare const createPostresEventStoreProvider: <T extends EventBase>({ eventSchema, poolConfig, }: Props) => EventStoreProvider<T>;
+export declare const createPostgresEventStoreProvider: <T extends EventBase>({ eventSchema, poolConfig, }: Props) => EventStoreProvider<T>;
 export {};
 //# sourceMappingURL=index.d.ts.map